Added cgiRedirectApClientToIP
Added cgiRedirectApClientToIP function.

If the client is on the SoftAP interface, and the hostname doesn’t
match the server’s IP address, the client gets redirected to the server
IP.

Used in conjunction with the captive dns function. For example, client
attempts connection to www.test.com, captive DNS says host IP is
192.168.4.1, client gets sent to esphttpd server, server detects
hostname is incorrect, redirects to http://192.168.4.1